+++ pkg/analyzer/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Remove deprecated `DartType.isStructurallyEqualTo`.
 * Remove deprecated `RecordType.positionalTypes`.
 * Remove deprecated `RecordType.sortedNamedTypes`.
+* Remove `ElementLocation` class, its values are not returned anymore.
 
 ## 7.4.1
 * Restore `InstanceElement.augmented` getter.